feat: StringUtils 에서 단순 String 처리가 아닌 것들 HttpUtils로 분리
